Extract metadata connection loader factories
